@@ -1,94 +1,44 @@
 {
-  "name": "material-dashboard-angular",
-  "version": "2.5.0",
-  "license": "MIT",
+  "name": "ui",
+  "version": "0.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e",
-    "install:clean": "rm -rf node_modules/ && rm -rf package-lock.json && npm install && ng serve -o"
-  },
-  "engines": {
-    "node": "6.11.1",
-    "npm": "3.10.9"
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "^1.1.0",
-    "@angular/animations": "11.2.3",
-    "@angular/cdk": "10.2.4",
-    "@angular/common": "^11.2.3",
-    "@angular/compiler": "11.2.3",
-    "@angular/core": "^11.2.3",
-    "@angular/forms": "11.2.3",
-    "@angular/http": "^7.2.16",
-    "@angular/localize": "^11.2.3",
-    "@angular/material": "10.2.4",
-    "@angular/platform-browser": "11.2.3",
-    "@angular/platform-browser-dynamic": "11.2.3",
-    "@angular/platform-server": "11.2.3",
-    "@angular/router": "11.2.3",
-    "@auth0/angular-jwt": "^5.0.2",
-    "@ng-bootstrap/ng-bootstrap": "^8.0.0",
-    "@ngx-translate/core": "^13.0.0",
-    "@ngx-translate/http-loader": "^6.0.0",
-    "@sweetalert2/ngx-sweetalert2": "^9.0.0",
-    "ajv": "6.12.5",
+    "@angular/animations": "^16.0.0",
+    "@angular/common": "^16.0.0",
+    "@angular/compiler": "^16.0.0",
+    "@angular/core": "^16.0.0",
+    "@angular/forms": "^16.0.0",
+    "@angular/platform-browser": "^16.0.0",
+    "@angular/platform-browser-dynamic": "^16.0.0",
+    "@angular/router": "^16.0.0",
+    "@auth0/angular-jwt": "^5.1.2",
+    "@gilsdav/ngx-translate-router": "^6.1.0",
+    "@ngx-translate/core": "^15.0.0",
+    "@ngx-translate/http-loader": "^8.0.0",
     "alertifyjs": "^1.13.1",
-    "arrive": "2.4.1",
-    "bootstrap": "4.5.2",
-    "bootstrap-material-design": "4.1.3",
-    "bootstrap-notify": "3.1.3",
-    "bootstrap-select": "^1.13.18",
-    "chartist": "0.11.4",
-    "classlist.js": "1.1.20150312",
-    "core-js": "3.6.5",
-    "datatables.net": "^1.10.20",
-    "datatables.net-dt": "^1.10.20",
-    "eslint": "^7.15.0",
-    "express": "4.17.1",
-    "font-awesome": "^4.7.0",
-    "googleapis": "61.0.0",
-    "hammerjs": "2.0.8",
-    "jquery": "3.5.1",
-    "moment": "2.29.1",
-    "ng-multiselect-dropdown": "^0.2.12",
-    "perfect-scrollbar": "1.5.0",
-    "popper.js": "^1.16.1",
-    "rxjs": "6.6.3",
-    "rxjs-compat": "6.6.3",
-    "sweetalert2": "^10.14.0",
-    "web-animations-js": "2.3.2",
-    "zone.js": "0.11.4"
+    "bootstrap": "^5.2.3",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1102.2",
-    "@angular/cli": "^11.2.2",
-    "@angular/compiler-cli": "11.2.3",
-    "@angular/language-service": "11.2.3",
-    "@fortawesome/fontawesome-free": "^5.15.1",
-    "@types/bootstrap": "4.5.0",
-    "@types/chartist": "0.11.0",
-    "@types/datatables.net": "^1.10.18",
-    "@types/googlemaps": "3.39.14",
-    "@types/jasmine": "3.5.14",
-    "@types/jquery": "3.5.2",
-    "@types/node": "14.11.5",
-    "codelyzer": "6.0.1",
-    "jasmine-core": "3.6.0",
-    "jasmine-spec-reporter": "6.0.0",
-    "karma": "5.2.3",
-    "karma-chrome-launcher": "3.1.0",
-    "karma-cli": "2.0.0",
-    "karma-coverage-istanbul-reporter": "3.0.3",
-    "karma-jasmine": "4.0.1",
-    "karma-jasmine-html-reporter": "1.5.4",
-    "protractor": "7.0.0",
-    "ts-node": "9.0.0",
-    "tslint": "6.1.3",
-    "typescript": "4.0.3"
+    "@angular-devkit/build-angular": "^16.0.1",
+    "@angular/cli": "~16.0.1",
+    "@angular/compiler-cli": "^16.0.0",
+    "@types/jasmine": "~4.3.0",
+    "jasmine-core": "~4.6.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.0.0",
+    "typescript": "~5.0.2"
   }
 }
